Move config, display out of noether.core
