+++ noether/__init__.py
@@ -1,2 +1,3 @@
+from .config import conf
 from .core import *
 from .catalogue import *
