+++ noether/catalogue/__init__.py
@@ -11,7 +11,7 @@ from .data import *
 from .conventional import *
 from .imperial import *
 
-from ..core.config import Config, conf
+from ..config import Config, conf
 
 Config.register('catalogue_extended', True, help='''
 Provide historical, cgs, obscure and regional units. Disabling these may provide some speedup.''')
