+++ noether/core/__init__.py
@@ -1,5 +1,3 @@
-from .config import Config, conf
-
 from .Dimension import Dimension
 from .Measure import Measure
 from .Unit import Unit
