+++ noether/display.py
